requirements.txt
# langchain
# openai
# huggingface_hub
# python-dotenv
# streamlit

langchain
openai
python-dotenv
streamlit